requirements.txt
fastapi==0.136.0
uvicorn==0.45.0
psycopg2-binary==2.9.11
bcrypt==5.0.0
python-multipart==0.0.5
